Added SceneAssistant and removed adding missiles and ufos from the core
